Move WorkerExt to a different file.
