+++ crates/worker/src/worker_ext.rs
@@ -1,0 +1,39 @@
+use js_sys::Uint8Array;
+use wasm_bindgen::closure::Closure;
+use wasm_bindgen::{JsCast, JsValue};
+use web_sys::{DedicatedWorkerGlobalScope, MessageEvent, Worker};
+
+pub(crate) fn worker_self() -> DedicatedWorkerGlobalScope {
+    JsValue::from(js_sys::global()).into()
+}
+
+pub(crate) trait WorkerExt {
+    fn set_onmessage_closure(&self, handler: impl 'static + Fn(Vec<u8>));
+
+    fn post_message_vec(&self, data: Vec<u8>);
+}
+
+macro_rules! worker_ext_impl {
+    ($($type:path),+) => {$(
+        impl WorkerExt for $type {
+            fn set_onmessage_closure(&self, handler: impl 'static + Fn(Vec<u8>)) {
+                let handler = move |message: MessageEvent| {
+                    let data = Uint8Array::from(message.data()).to_vec();
+                    handler(data);
+                };
+                let closure = Closure::wrap(Box::new(handler) as Box<dyn Fn(MessageEvent)>);
+                self.set_onmessage(Some(closure.as_ref().unchecked_ref()));
+                closure.forget();
+            }
+
+            fn post_message_vec(&self, data: Vec<u8>) {
+                self.post_message(&Uint8Array::from(data.as_slice()))
+                    .expect("failed to post message");
+            }
+        }
+    )+};
+}
+
+worker_ext_impl! {
+    Worker, DedicatedWorkerGlobalScope
+}
